[Add] #1 - 프로젝트에 필요한 기본 파일 및 폰트 추가
