+++ 2024-NC2-A31-WhereAreYou/-024-NC2-A31-WhereAreYou-Info.plist
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+	<key>UIAppFonts</key>
+	<array>
+		<string>Galmuri7.ttf</string>
+		<string>Galmuri9.ttf</string>
+	</array>
+</dict>
+</plist>
