+++ 2024-NC2-A31-WhereAreYou/2024-NC2-A31-WhereAreYou/App/_024_NC2_A31_WhereAreYouApp.swift
@@ -1,0 +1,17 @@
+//
+//  _024_NC2_A31_WhereAreYouApp.swift
+//  2024-NC2-A31-WhereAreYou
+//
+//  Created by Seoyeon Choi on 6/16/24.
+//
+
+import SwiftUI
+
+@main
+struct _024_NC2_A31_WhereAreYouApp: App {
+    var body: some Scene {
+        WindowGroup {
+            ContentView()
+        }
+    }
+}
