--- 2024-NC2-A31-WhereAreYou/2024-NC2-A31-WhereAreYou/ContentView.swift
@@ -1,24 +1,0 @@
-//
-//  ContentView.swift
-//  2024-NC2-A31-WhereAreYou
-//
-//  Created by Seoyeon Choi on 6/16/24.
-//
-
-import SwiftUI
-
-struct ContentView: View {
-    var body: some View {
-        VStack {
-            Image(systemName: "heart.fill")
-                .imageScale(.large)
-                .foregroundStyle(.pink)
-            Text("구블구블")
-        }
-        .padding()
-    }
-}
-
-#Preview {
-    ContentView()
-}
