+++ 2024-NC2-A31-WhereAreYou/2024-NC2-A31-WhereAreYou/Extensions/Font+.swift
@@ -1,0 +1,14 @@
+//
+//  Font+.swift
+//  2024-NC2-A31-WhereAreYou
+//
+//  Created by Seoyeon Choi on 6/17/24.
+//
+
+import SwiftUI
+
+extension Font {
+    // 추후 수정 필요
+    static let galmuri7: Font = .custom("Galmuri7", size: 24)
+    static let galmuri9: Font = .custom("Galmuri9", size: 20)
+}
