+++ 2024-NC2-A31-WhereAreYou/2024-NC2-A31-WhereAreYou/Resources/StringLiterals.swift
@@ -1,0 +1,13 @@
+//
+//  StringLiterals.swift
+//  2024-NC2-A31-WhereAreYou
+//
+//  Created by Seoyeon Choi on 6/17/24.
+//
+
+import Foundation
+
+enum StringLiterals {
+    // 앱에 필요한 고정 String은 여기에 추가해서 사용하기!
+    static let mainTitle = "어디세요?"
+}
