+++ 2024-NC2-A31-WhereAreYou/2024-NC2-A31-WhereAreYou/Views/ChannelDetail/ChannelDetailView.swift
@@ -1,0 +1,19 @@
+//
+//  ChannelDetailView.swift
+//  2024-NC2-A31-WhereAreYou
+//
+//  Created by Seoyeon Choi on 6/17/24.
+//
+
+import SwiftUI
+
+struct ChannelDetailView: View {
+    var body: some View {
+        Text("ChannelDetailView")
+            .font(.galmuri7)
+    }
+}
+
+#Preview {
+    ChannelDetailView()
+}
