+++ 2024-NC2-A31-WhereAreYou/2024-NC2-A31-WhereAreYou/Views/ChannelList/ChannelListView.swift
@@ -1,0 +1,19 @@
+//
+//  ChannelListView.swift
+//  2024-NC2-A31-WhereAreYou
+//
+//  Created by Seoyeon Choi on 6/17/24.
+//
+
+import SwiftUI
+
+struct ChannelListView: View {
+    var body: some View {
+        Text("ChannelListView")
+            .font(.galmuri9)
+    }
+}
+
+#Preview {
+    ChannelListView()
+}
