+++ 2024-NC2-A31-WhereAreYou/2024-NC2-A31-WhereAreYou/Views/ChannelList/NewChannelView.swift
@@ -1,0 +1,19 @@
+//
+//  NewChannelView.swift
+//  2024-NC2-A31-WhereAreYou
+//
+//  Created by Seoyeon Choi on 6/17/24.
+//
+
+import SwiftUI
+
+struct NewChannelView: View {
+    var body: some View {
+        Text("NewChannelView")
+            .font(.galmuri7)
+    }
+}
+
+#Preview {
+    NewChannelView()
+}
